backend/requirements.txt
Flask-Cors==4.0.1
Flask==3.0.3
sentence-transformers==3.0.0
redis==5.0.4
elasticsearch==8.13.2
python-dotenv==1.0.1
feedparser==6.0.11
transformers==4.41.1
datasets==2.20.0
mypy==1.10.0
openai==1.35.3
torch==2.3.0
spacy==3.7.5